Implementation of data validation

+++ SklepAPI/Models/ProductsGetUserOrdersDto.cs
@@ -4,7 +4,7 @@
     {
         public int ProductId { get; set; }
         public string ProductName { get; set; }
-        public int ProductPrice { get; set; }
+        public double ProductPrice { get; set; }
         public int Quantity { get; set; }
     }
 }
